the first part of the application (getUserData) is ready and works correctly

+++ src/main/kotlin/com/persolog/halis/Repositories.kt
@@ -8,5 +8,5 @@ interface ModeRepository: CrudRepository<Mode, Byte> {
 interface RoomRepository: CrudRepository<Room, Byte> {
 }
 
-interface DataRepository: CrudRepository<Data, Byte> {
+interface DataRepository: CrudRepository<Data, Long> {
 }
